refactor: Core import-export structure
