+++ src/config/index.ts
@@ -1,0 +1,22 @@
+import crud from './crud'
+import dbClient from './db'
+//import * as output from './output'
+import {
+    moduleName,
+    pathBackend,
+    pathFrontend, 
+    pathModule, 
+    pathUser
+} from './output'
+
+export {
+    crud,
+    dbClient,
+    //output
+    moduleName,
+    pathBackend,
+    pathFrontend, 
+    pathModule, 
+    pathUser
+}
+
