+++ src/stack/index.ts
@@ -1,0 +1,4 @@
+import backendProcess from '@stackProcesses/laravel'
+import frontendProcess from '@stackProcesses/vue'
+
+export { backendProcess, frontendProcess }
