+++ tsconfig.json
@@ -31,6 +31,7 @@
     "baseUrl": "./src",                                  /* Specify the base directory to resolve non-relative module names. */
     "paths": {
       "@config/*": ["config/*"],
+      "@core/*": ["core/*"],
       "@coreModules/*": ["core/modules/*"],
       "@coreProcesses/*": ["core/processes/*"],
       "@coreTypes/*": ["core/types/*"],
